Reorganized card into multiple components
